Remove default values check test
